First implementation of Advanced Custom Fields.
It is now possible to add local JSON files for ACF in the
acf folder. These files are automatically read and will generate
a meta-box in the post-edit page; the choice made by the post
author will be sent to Google Analytics as a contentGroup.

See ACF documentation for more details:
www.advancedcustomfields.com/resources/local-json/.

In the next commits I will do the following:
1) Allow the user to choose which ACF field group to use from
those already installed in ACF.
2) In alternative, allow the users to upload their own JSON
files via the settings page.