Ignore the top image in scroll-tracking algorithm.
If a webpage's content starts with an image, the user is likely to scroll
past that image in order to reach the text below. This scrolling should
not be considered as engagement.

Wordpress Analytics now ignores this scrolling if the content is a Blog
entry, by ignoring any image within the first paragraph of a "div.entry-content"
element.

We plan to extend this feature also to the other types of content: recipes,
products and uncategorized.